mcp-erp-demo/product-python-rest-crud-postgres/run.py
import uvicorn
import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

if __name__ == "__main__":
    host = os.getenv("API_HOST", "0.0.0.0")
    port = int(os.getenv("API_PORT", "8000"))
    debug = os.getenv("DEBUG", "False").lower() in ("true", "1", "t")
    
    print(f"Starting ERP Product Catalog API on http://{host}:{port}")
    uvicorn.run("app.main:app", host=host, port=port, reload=debug)
